+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_07_26_012622) do
+ActiveRecord::Schema.define(version: 2020_08_11_194846) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -150,6 +150,7 @@ ActiveRecord::Schema.define(version: 2020_07_26_012622) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.bigint "assessment_id", null: false
+    t.boolean "bonus", default: false, null: false
     t.index ["assessment_id"], name: "index_criteria_on_assessment_id"
   end